backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pip==24.0
psycopg2==2.9.9
setuptools==69.1.0
wheel==0.42.0
